requirements.txt
selenium
SQLAlchemy
libmysqlclient-dev
mysqlclient